Refactors Meta Header
- Extracts meta header related code into a module
- Adds more tests for specific situations when http libraries haven't been loaded by the time the meta header methods are called

Related to #1224
